roulette_backend/requirements.txt
asgiref==3.7.2
black==24.2.0
click==8.1.7
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
gunicorn==21.2.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
